+++ lib/src/tooltip_widget.dart
@@ -29,6 +29,7 @@ import 'get_position.dart';
 import 'measure_size.dart';
 import 'models/tooltip_action_config.dart';
 import 'widget/action_widget.dart';
+import 'widget/floating_action_widget.dart';
 import 'widget/tooltip_slide_transition.dart';
 
 class ToolTipWidget extends StatefulWidget {
@@ -44,7 +45,7 @@ class ToolTipWidget extends StatefulWidget {
   final TextStyle? titleTextStyle;
   final TextStyle? descTextStyle;
   final Widget? container;
-  final Widget? staticContainer;
+  final FloatingActionWidget? floatingActionWidget;
   final Color? tooltipBackgroundColor;
   final Color? textColor;
   final bool showArrow;
@@ -80,7 +81,7 @@ class ToolTipWidget extends StatefulWidget {
     required this.titleTextStyle,
     required this.descTextStyle,
     required this.container,
-    required this.staticContainer,
+    required this.floatingActionWidget,
     required this.tooltipBackgroundColor,
     required this.textColor,
     required this.showArrow,
@@ -472,7 +473,7 @@ class _ToolTipWidgetState extends State<ToolTipWidget>
     }
 
     if (widget.container == null) {
-      return Positioned(
+      final defaultToolTipWidget = Positioned(
         top: contentY,
         left: _getLeft(),
         right: _getRight(),
@@ -653,9 +654,22 @@ class _ToolTipWidgetState extends State<ToolTipWidget>
           ),
         ),
       );
+
+      if (widget.floatingActionWidget != null) {
+        return Stack(
+          fit: StackFit.expand,
+          children: [
+            defaultToolTipWidget,
+            widget.floatingActionWidget!,
+          ],
+        );
+      } else {
+        return defaultToolTipWidget;
+      }
     }
 
     return Stack(
+      fit: StackFit.expand,
       children: <Widget>[
         Positioned(
           left: _getSpace(),
@@ -723,7 +737,7 @@ class _ToolTipWidgetState extends State<ToolTipWidget>
             ),
           ),
         ),
-        if (widget.staticContainer != null) ...[widget.staticContainer!],
+        if (widget.floatingActionWidget != null) widget.floatingActionWidget!,
       ],
     );
   }